+++ src/index.ts
@@ -1,6 +1,7 @@
 export * from "./UserProvider.js";
 export * from "./SettingsProvider.js";
 export * from "./accessibility.js";
+export * from "./i18n.js";
 export * from "./profile-save.js";
 export * from "./Pages/Settings/index.js";
 export * from "./components/avatar-upload-panel/index.js";